+++ groom_macros/tests/expand/01-response.expanded.rs
@@ -3032,7 +3032,10 @@ mod result_struct_struct {
     }
     pub fn merge_into_router(
         other: impl Into<::groom::router::GroomRouter<()>>,
-    ) -> ::groom::router::GroomRouter<()> {
+    ) -> ::std::result::Result<
+        ::groom::router::GroomRouter<()>,
+        ::groom::router::MergeError,
+    > {
         __groom_runtime_checks();
         let this_router: ::axum::Router<()> = ::axum::Router::new()
             .route("/", ::axum::routing::get(__groom_wrapper_foo));
@@ -3066,14 +3069,7 @@ mod result_struct_struct {
             __groom_paths,
         );
         let __groom_other = other.into();
-        match __groom_other.merge(__groom_this) {
-            ::std::result::Result::Ok(r) => r,
-            ::std::result::Result::Err(e) => {
-                ::core::panicking::panic_fmt(
-                    format_args!("GroomRouter merge failed: {0}", e),
-                );
-            }
-        }
+        __groom_other.merge(__groom_this)
     }
     const _: fn() = || {
         fn assert_impl_all<T: ?Sized + ::groom::response::Response>() {}
@@ -3755,7 +3751,10 @@ mod result_struct_enum {
     }
     pub fn merge_into_router(
         other: impl Into<::groom::router::GroomRouter<()>>,
-    ) -> ::groom::router::GroomRouter<()> {
+    ) -> ::std::result::Result<
+        ::groom::router::GroomRouter<()>,
+        ::groom::router::MergeError,
+    > {
         __groom_runtime_checks();
         let this_router: ::axum::Router<()> = ::axum::Router::new()
             .route("/", ::axum::routing::get(__groom_wrapper_foo));
@@ -3789,14 +3788,7 @@ mod result_struct_enum {
             __groom_paths,
         );
         let __groom_other = other.into();
-        match __groom_other.merge(__groom_this) {
-            ::std::result::Result::Ok(r) => r,
-            ::std::result::Result::Err(e) => {
-                ::core::panicking::panic_fmt(
-                    format_args!("GroomRouter merge failed: {0}", e),
-                );
-            }
-        }
+        __groom_other.merge(__groom_this)
     }
     const _: fn() = || {
         fn assert_impl_all<T: ?Sized + ::groom::response::Response>() {}
@@ -4642,7 +4634,10 @@ mod wrapped_enum {
     }
     pub fn merge_into_router(
         other: impl Into<::groom::router::GroomRouter<()>>,
-    ) -> ::groom::router::GroomRouter<()> {
+    ) -> ::std::result::Result<
+        ::groom::router::GroomRouter<()>,
+        ::groom::router::MergeError,
+    > {
         __groom_runtime_checks();
         let this_router: ::axum::Router<()> = ::axum::Router::new()
             .route("/", ::axum::routing::get(__groom_wrapper_foo));
@@ -4676,14 +4671,7 @@ mod wrapped_enum {
             __groom_paths,
         );
         let __groom_other = other.into();
-        match __groom_other.merge(__groom_this) {
-            ::std::result::Result::Ok(r) => r,
-            ::std::result::Result::Err(e) => {
-                ::core::panicking::panic_fmt(
-                    format_args!("GroomRouter merge failed: {0}", e),
-                );
-            }
-        }
+        __groom_other.merge(__groom_this)
     }
     const _: fn() = || {
         fn assert_impl_all<T: ?Sized + ::groom::response::Response>() {}